move enums to defs
