--- src/core/action_enum.py
@@ -1,8 +1,0 @@
-from enum import Enum
-
-
-class AgentAction(Enum):
-    IDLE = "idle"
-    CHECK_NEWS = "check_news"
-    POST_TO_TELEGRAM = "post_to_telegram"
-    POST_TO_TWITTER = "post_to_twitter"
