+++ src/core/defs.py
@@ -9,3 +9,22 @@ class Environment(Enum):
     PRODUCTION = "production"
     DEVELOPMENT = "development"
     CI = "ci"
+
+
+class AgentAction(Enum):
+    """The actions that the agent can take."""
+
+    IDLE = "idle"
+    CHECK_NEWS = "check_news"
+    POST_TO_TELEGRAM = "post_to_telegram"
+    POST_TO_TWITTER = "post_to_twitter"
+
+
+class AgentState(Enum):
+    """The states that the agent can be in."""
+
+    DEFAULT = "default"
+    IDLE = "idle"
+    WAITING_FOR_NEWS = "waiting_for_news"
+    JUST_POSTED_TO_TELEGRAM = "just_posted_to_telegram"
+    JUST_POSTED_TO_TWITTER = "just_posted_to_twitter"
